Google tag code should work in non-client environments
When trying to render on the server side, the window object is not available and so the build will fail. This patch replaces window with null if it is undefined and then has the tag include code return immediately if so. This allows us to build server side, while keeping the google tag code running.